Add files via upload
The paradigm shift in technology which has enabled corporations to handle copious amount of data
and extract meaningful information has increased the competition among various financial
corporations which has made creation of data driven business and marketing strategy a crucial part
of Banking sector. Following the same path, the business case was to apply data analytics to make
empower strategies based on targeted marketing and focus on customers which would have been
possible through understanding the transactional behavior of customers and reckoning similarities
among customers for grouping. By focusing on the business problem and unsupervised transactional
data from DataMart the solution was to impose unsupervised learning and apply clustering on the
dataset. The dataset is comprised of both categorical and numerical data with an ample number of
uncertain columns and thus different types of clustering methods and machine learning techniques
along with feature engineering have been used.
Clustering is a type of unsupervised learning which involves grouping a set of objects in such a
the way that objects in the same group are more like each other than those in other groups. Data
mining, EDA, and different types of clustering models have been used to perceive the best clusters
substantiated by Data mining, EDA, and cluster visualizations to understand the clusters and infer
business outcome